fixing the repo also the bullets should now be updated to cause ragdoll and splatter
